refactor(settings): extract agent list fields editor seam
Move the editor, field editor, and shared error view into named files,
add dirtyHostIDs and underlyingSource, register a Console-card preview
seam, and add simulator-id make targets.

refs #281
