+++ Sources/Heeler/Settings/AgentListFieldsSettingsView.swift
@@ -1,161 +1,4 @@
 import SwiftUI
-import Observation
-
-/// Read-only until Edit. Every change, including a Sync from plugin fill,
-/// lands in a per-Host draft that only Save persists; Cancel and leaving the
-/// screen discard drafts. Hosts without a saved choice show their herdr
-/// fields, or Heeler's fallback when the plugin has none.
-@MainActor
-@Observable
-final class AgentListFieldsEditor {
-    enum SyncState: Equatable {
-        case syncing
-        case filled(String)
-        case failed(String)
-
-        var message: String? {
-            switch self {
-            case .syncing: nil
-            case .filled(let message), .failed(let message): message
-            }
-        }
-    }
-
-    enum LayoutSource: Equatable {
-        case draft, saved, plugin, pluginDefaults, loading, missing, unavailable
-    }
-
-    typealias Fetch = @MainActor (Host.ID) async -> HerdrSidebarSnapshotStore.HostState?
-
-    let layouts: AgentRowLayoutStore
-    let snapshots: HerdrSidebarSnapshotStore
-    private let fetch: Fetch
-    private(set) var isEditing = false
-    private(set) var drafts: [Host.ID: AgentRowLayout] = [:]
-    private(set) var syncStates: [Host.ID: SyncState] = [:]
-    var errorMessage: String?
-    @ObservationIgnored private var syncRequests: [Host.ID: UUID] = [:]
-
-    init(layouts: AgentRowLayoutStore, snapshots: HerdrSidebarSnapshotStore, fetch: @escaping Fetch) {
-        self.layouts = layouts
-        self.snapshots = snapshots
-        self.fetch = fetch
-    }
-
-    /// The Console's layout for this Host, or its draft while editing.
-    func layout(for hostID: Host.ID) -> AgentRowLayout {
-        if isEditing, let draft = drafts[hostID] { return draft }
-        return layouts.resolvedLayout(for: hostID, pluginSnapshot: snapshots.snapshot(for: hostID))
-    }
-
-    func source(for hostID: Host.ID) -> LayoutSource {
-        if isEditing, drafts[hostID] != nil { return .draft }
-        if layouts.hostLayouts[hostID] != nil { return .saved }
-        switch snapshots.states[hostID] {
-        case .loading: return .loading
-        case .loaded(let snapshot?): return snapshot.diagnostics.isEmpty ? .plugin : .pluginDefaults
-        case .loaded(nil): return .missing
-        case .unavailable, nil: return .unavailable
-        }
-    }
-
-    /// True when Save would change the persisted catalog.
-    var hasUnsavedChanges: Bool {
-        drafts.contains { $0.value != layouts.hostLayouts[$0.key] }
-    }
-
-    func beginEditing() {
-        isEditing = true
-        clearDrafts()
-    }
-
-    func cancel() {
-        isEditing = false
-        clearDrafts()
-    }
-
-    /// All changed Hosts are written in one validated step; a failed write
-    /// keeps the drafts and stays in edit mode.
-    func save() {
-        guard isEditing else { return }
-        let changes = drafts.filter { $0.value != layouts.hostLayouts[$0.key] }
-            .mapValues { Optional($0) }
-        do {
-            try layouts.setLayouts(changes)
-            errorMessage = nil
-            isEditing = false
-            clearDrafts()
-        } catch {
-            report(error)
-        }
-    }
-
-    func update(_ hostID: Host.ID, _ edit: (inout AgentRowLayout) -> Void) {
-        guard isEditing else { return }
-        var next = layout(for: hostID)
-        edit(&next)
-        do {
-            try next.validate()
-        } catch {
-            report(error)
-            return
-        }
-        drafts[hostID] = next
-        // A later edit wins over an in-flight or finished sync.
-        syncRequests[hostID] = nil
-        syncStates[hostID] = nil
-        errorMessage = nil
-    }
-
-    func setRows(_ rows: [AgentRow], kind: String?, for hostID: Host.ID) {
-        update(hostID) {
-            if let kind { $0.rowsByAgent[kind] = rows }
-            else { $0.rows = rows }
-        }
-    }
-
-    /// Fetches the Host's plugin snapshot and fills its draft. A missing
-    /// snapshot fills Heeler's fallback fields; a failed read changes nothing.
-    /// Results are dropped once editing ended or the draft was edited since.
-    func syncFromPlugin(_ hostID: Host.ID) async {
-        guard isEditing else { return }
-        let request = UUID()
-        syncRequests[hostID] = request
-        syncStates[hostID] = .syncing
-        let state = await fetch(hostID)
-        guard isEditing, syncRequests[hostID] == request else { return }
-        syncRequests[hostID] = nil
-        switch state {
-        case .loaded(let snapshot?):
-            drafts[hostID] = snapshot.layout
-            syncStates[hostID] = .filled(snapshot.diagnostics.isEmpty
-                ? "Filled from the herdr plugin. Save to keep these fields."
-                : "herdr reported a configuration problem, so its default fields were filled. Save to keep them.")
-        case .loaded(nil):
-            drafts[hostID] = .heelerDefault
-            syncStates[hostID] = .filled(
-                "The herdr plugin on this Host has no fields snapshot, so Heeler's fallback fields were filled. Save to keep them.")
-        case .unavailable:
-            syncStates[hostID] = .failed("herdr fields could not be read from this Host. Nothing was changed.")
-        case .loading, nil:
-            syncStates[hostID] = .failed("This Host is not connected. Nothing was changed.")
-        }
-        errorMessage = nil
-    }
-
-    private func clearDrafts() {
-        drafts = [:]
-        syncStates = [:]
-        syncRequests = [:]
-        errorMessage = nil
-    }
-
-    private func report(_ error: any Error) {
-        errorMessage = error is AgentRowLayoutStoreError
-            ? "The saved Agent List Fields could not be read. Nothing was changed."
-            : "This layout could not be saved. Use at most 16 rows and 16 fields per row, with valid field names."
-    }
-}
 
 struct AgentListFieldsSettingsView: View {
     let console: ConsoleStore
@@ -496,98 +339,6 @@ private struct AgentLayoutRowsContent: View {
                 Text(verbatim: rowMetadata(rows[index]))
                     .font(.caption2)
                     .foregroundStyle(.tertiary)
-            }
-        }
-    }
-}
-
-private struct AgentLayoutTokensView: View {
-    let editor: AgentListFieldsEditor
-    let hostID: Host.ID
-    let kind: String?
-    let rowIndex: Int
-    @State private var customName = ""
-
-    private var rows: [AgentRow] {
-        let layout = editor.layout(for: hostID)
-        return kind.map { layout.rowsByAgent[$0] ?? [] } ?? layout.rows
-    }
-    private var tokens: AgentRow { rows.indices.contains(rowIndex) ? rows[rowIndex] : [] }
-
-    var body: some View {
-        List {
-            Section {
-                ForEach(Array(tokens.indices), id: \.self) { index in
-                    Text(verbatim: tokens[index].token.rawValue)
-                }
-                .onDelete(perform: deleteHandler)
-                .onMove(perform: moveHandler)
-            } header: {
-                Text("Fields")
-            } footer: {
-                Text("Status fields are accepted but shown in the status column. Edit to delete or reorder fields.")
-            }
-            if editor.isEditing {
-                Section {
-                    Menu("Add Field", systemImage: "plus") {
-                        ForEach(AgentRowToken.builtins, id: \.self) { token in
-                            Button(token.rawValue) { mutate { $0.append(.init(token)) } }
-                        }
-                    }
-                    TextField("$custom_name", text: $customName)
-                        .textInputAutocapitalization(.never)
-                        .autocorrectionDisabled()
-                    Button("Add Custom Field") {
-                        guard let token = customToken else { return }
-                        mutate { $0.append(.init(token)) }
-                        if editor.errorMessage == nil { customName = "" }
-                    }
-                    .disabled(customToken == nil)
-                } footer: {
-                    Text("Custom names start with $ and contain 1–32 letters, digits, underscores or hyphens. Values come from herdr plugins and display as plain text.")
-                }
-                .disabled(tokens.count >= AgentRowLayout.maximumTokensPerRow || !rows.indices.contains(rowIndex))
-            }
-            AgentLayoutErrorView(editor: editor)
-        }
-        .navigationTitle("Row \(rowIndex + 1)")
-        .toolbar {
-            if editor.isEditing { EditButton() }
-        }
-    }
-
-    private var customToken: AgentRowToken? {
-        guard customName.hasPrefix("$") else { return nil }
-        return AgentRowToken(rawValue: customName)
-    }
-
-    private var deleteHandler: ((IndexSet) -> Void)? {
-        guard editor.isEditing else { return nil }
-        return { offsets in mutate { $0.remove(atOffsets: offsets) } }
-    }
-
-    private var moveHandler: ((IndexSet, Int) -> Void)? {
-        guard editor.isEditing else { return nil }
-        return { offsets, destination in mutate { $0.move(fromOffsets: offsets, toOffset: destination) } }
-    }
-
-    private func mutate(_ change: (inout AgentRow) -> Void) {
-        var next = rows
-        guard next.indices.contains(rowIndex) else { return }
-        change(&next[rowIndex])
-        editor.setRows(next, kind: kind, for: hostID)
-    }
-}
-
-private struct AgentLayoutErrorView: View {
-    let editor: AgentListFieldsEditor
-
-    var body: some View {
-        if let message = editor.errorMessage {
-            Section {
-                Text(verbatim: message)
-                    .foregroundStyle(.red)
-                    .accessibilityIdentifier("settings.agentList.error")
             }
         }
     }
