+++ Tests/HeelerTests/AgentLayoutTokensViewTests.swift
@@ -1,0 +1,6 @@
+import Testing
+
+@testable import Heeler
+
+@Suite("Agent layout tokens view")
+struct AgentLayoutTokensViewTests {}
