+++ Tests/HeelerTests/AgentListFieldsPreviewTests.swift
@@ -1,0 +1,6 @@
+import Testing
+
+@testable import Heeler
+
+@Suite("Agent list fields preview")
+struct AgentListFieldsPreviewTests {}
